requirements.txt
fastapi
uvicorn
pydantic
motor  
pytz   
bcrypt 
python-dotenv 
pydantic[email]




